fixed the response returning object in the auth.js.

+++ backend/routes/notes.js
@@ -13,13 +13,14 @@ const new_note_validations = [
 router.get("/fetchallnotes", fetchuer, async (req, res) => {
     try {
         // getting notes 
-        const notes = await Notes.find({ user: req.id });//user is already appended in the req using fetchuser middleware
+        const notes = await Notes.find({ user: req.user.id });//user is already appended in the req using fetchuser middleware
+        // console.log(req.user)
         // sending notes as response
         res.json(notes)
     } catch (error) {
 
         res.status(500).json({ "error": "some problem occured..."})
-        console.log(error)
+        // console.log(error)
     }
 
 })
@@ -36,7 +37,7 @@ router.post("/addnewnote", new_note_validations, fetchuer, async (req, res) => {
 
         // creating notes using object desctructuring
         const note = await new Notes({
-            title, description, tag, user: req.id
+            title, description, tag, user: req.user.id
         })
         const saved_notes = await note.save()
         res.json(saved_notes)
@@ -58,9 +59,9 @@ router.put("/updatenote/:noteid", fetchuer, async (req, res) => {
         if (!note) { return res.status(404).send("Not found") };
 
         // authorizing user id matching with the user id in the notes 
-        // if (note.user.toString() !== req.id) {
-        //     return res.status(401).send("Not Allowed")
-        // }
+        if (note.user.toString() !== req.user.id) {
+            return res.status(401).send("Not Allowed")
+        }
 
         // find the note to be updated
         const create_new_note = await Notes.findByIdAndUpdate(req.params.noteid, { $set: newNote }, { $new: true });
@@ -79,14 +80,14 @@ router.delete("/deletenote/:noteid", fetchuer, async (req, res) => {
         if (!note) { return res.status(404).send("Not found") };
 
         // authorizing user id matching with the user id in the notes
-        // if (note.user.toString() !== req.id) {
-        //     return res.status(401).send("Not Allowed")
-        // }
+        if (note.user.toString() !== req.user.id) {
+            return res.status(401).send("Not Allowed")
+        }
 
         await Notes.findByIdAndDelete(req.params.noteid);
         res.json({ Status: "Delete Success", note: note })
     } catch (error) {
-        console.log(error)
+        // console.log(error)
         res.status(500).json({ "error": "some problem occured..." })
     }
 })
